Move IDEA proxy to demo main script
